requirements.txt
streamlit
pandas
torch
torchvision
transformers
Pillow
requests
numpy
